requirements.txt
aiogram==2.25.1
pillow==7.0.0
